requirements/dev.txt
mock
coverage
flake8
tox
pytest
pytest-django
pytest-cov
docutils
python-creole
wheel
twine
aldryn_translation_tools

# https://github.com/divio/django-cms
django-cms

django-parler
django-tools>=0.38.0
django-cms-tools>=0.5.0

# http://django-debug-toolbar.readthedocs.io/en/stable/changes.html
django-debug-toolbar
